Moved handling of /api/sensorEvent to sensorAlerts.js
