+++ server/test/unit/sensorAlertsTest/handleAlertTest.js
@@ -13,6 +13,9 @@ const db = require('../../../db/db')
 
 const sensorAlerts = rewire('../../../sensorAlerts')
 
+// eslint-disable-next-line no-underscore-dangle
+const handleAlert = sensorAlerts.__get__('handleAlert')
+
 let braveAlerter
 
 use(sinonChai)
@@ -52,7 +55,7 @@ describe('sensorAlerts.js unit tests: handleAlert', () => {
       sandbox.stub(db, 'createSession').returns(this.session)
       sandbox.stub(db, 'saveSession')
 
-      await sensorAlerts.handleAlert(this.location, alertType)
+      await handleAlert(this.location, alertType)
     })
 
     it('should log the alert', () => {
@@ -123,7 +126,7 @@ describe('sensorAlerts.js unit tests: handleAlert', () => {
       sandbox.stub(db, 'createSession').returns(this.session)
       sandbox.stub(db, 'saveSession')
 
-      await sensorAlerts.handleAlert(this.location, alertType)
+      await handleAlert(this.location, alertType)
     })
 
     it('should log the alert', () => {
@@ -195,7 +198,7 @@ describe('sensorAlerts.js unit tests: handleAlert', () => {
       sandbox.stub(db, 'createSession')
       sandbox.stub(db, 'saveSession')
 
-      await sensorAlerts.handleAlert(this.location, alertType)
+      await handleAlert(this.location, alertType)
     })
 
     it('should log the alert', () => {
@@ -248,7 +251,7 @@ describe('sensorAlerts.js unit tests: handleAlert', () => {
       sandbox.stub(db, 'createSession')
       sandbox.stub(db, 'saveSession')
 
-      await sensorAlerts.handleAlert(this.location, alertType)
+      await handleAlert(this.location, alertType)
     })
 
     it('should log the alert', () => {
@@ -301,7 +304,7 @@ describe('sensorAlerts.js unit tests: handleAlert', () => {
       sandbox.stub(db, 'createSession')
       sandbox.stub(db, 'saveSession')
 
-      await sensorAlerts.handleAlert(this.location, alertType)
+      await handleAlert(this.location, alertType)
     })
 
     it('should log the alert', () => {
